+++ tealish/errors.py
@@ -1,0 +1,7 @@
+
+class ParseError(Exception):
+    pass
+
+
+class CompileError(Exception):
+    pass
